# fix: use same UNC path normalization logic with libuv