--- a/Cargo.toml
+++ b/Cargo.toml
@@ -70,7 +70,6 @@ serde_json = { version = "1", features = [
   "preserve_order",
 ] } # preserve_order: package_json.exports requires order such as `["require", "import", "default"]`
 rustc-hash = { version = "2" }
-dunce = "1" # Normalize Windows paths to the most compatible format, avoiding UNC where possible
 once_cell = "1" # Use `std::sync::OnceLock::get_or_try_init` when it is stable.
 thiserror = "1"
 json-strip-comments = "1"