--- a/Cargo.toml
+++ b/Cargo.toml
@@ -14,7 +14,7 @@ keywords     = ["node", "resolve", "cjs", "esm", "enhanced-resolve"]
 license      = "MIT"
 readme       = "README.md"
 repository   = "https://github.com/oxc-project/oxc-resolver"
-rust-version = "1.70"
+rust-version = "1.74"
 include      = ["/src", "/examples", "/benches"]
 
 [lib]