--- a/fixtures/pnpm/longfilename/index.js
+++ b/fixtures/pnpm/longfilename/index.js
@@ -1,0 +1,3 @@
+const test = 'hello world'
+
+export default test