fixes #295