@Brooooooklyn 2024-11-13
@sapphi-red Please wait a moment while I verify if this custom canonicalize implementation is truly optimized on Windows