@sapphi-red 2024-11-13
@Brooooooklyn It seems `cargo bench` is slower on my machine.
Before
```
resolver/single-thread  time:   [143.72 µs 148.36 µs 154.09 µs]
Found 8 outliers among 100 measurements (8.00%)
  2 (2.00%) high mild
  6 (6.00%) high severe
resolver/multi-thread   time:   [117.85 µs 119.70 µs 121.46 µs]
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe
Benchmarking resolver/resolve from symlinks: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 301.3s, or reduce sample count to 10.
resolver/resolve from symlinks
                        time:   [2.9440 s 2.9456 s 2.9473 s]
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
```
After
```
resolver/single-thread  time:   [139.14 µs 144.00 µs 151.83 µs]
                        change: [−20.919% −3.8025% +16.565%] (p = 0.71 > 0.05)
                        No change in performance detected.
Found 12 outliers among 100 measurements (12.00%)
  4 (4.00%) high mild
  8 (8.00%) high severe
resolver/multi-thread   time:   [112.48 µs 114.09 µs 115.69 µs]
                        change: [−7.9616% −2.9621% +2.4978%] (p = 0.29 > 0.05)
                        No change in performance detected.
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) high mild
  4 (4.00%) high severe
Benchmarking resolver/resolve from symlinks: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 743.3s, or reduce sample count to 10.
resolver/resolve from symlinks
                        time:   [7.4256 s 7.4356 s 7.4479 s]
                        change: [+152.08% +152.43% +152.85%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) high mild
  4 (4.00%) high severe
```