@Brooooooklyn 2024-11-13
> it seems cargo bench is slower on my machine.

Me too, It seems that we can simply use `node_compatible_raw_canonicalize(fs::canonicalize(&path))` under Windows.