@sapphi-red 2024-11-13
I reverted some code to use `fs::canonicalize` for Windows.